requirement.txt
openai>=1.30.0
langchain>=0.2.0
langchain-community>=0.2.0
chromadb>=0.5.0
python-docx>=1.1.0
fastapi>=0.110.0
uvicorn>=0.29.0
python-multipart>=0.0.9
jieba>=0.42.1
